move availableFeeOptions to the estimation controller
